added StatsMonitor to track non-time data for measurement with statistics
